requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
openai>=1.3.0
jinja2>=3.1.0 
google-generativeai 